feat(TMW-000): Newsletter query provider
