+++ .prettierignore
@@ -18,7 +18,7 @@ storybook-static
 *.html
 *.sh
 *.log
-packages
+
 coverage
 settings.gradle
 build.gradle
